Specular lighting added with material class
Also changed window size, commented out scale transformation, and updated UseLight location parameters to GLuint

+++ Light.cpp
@@ -25,8 +25,8 @@ Light::Light(GLfloat red, GLfloat green, GLfloat blue, GLfloat aIntensity,
 	diffuseIntensity = dIntensity;
 }
 
-void Light::UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColorLocation,
-	                 GLfloat diffuseIntensityLocation, GLfloat directionLocation)
+void Light::UseLight(GLuint ambientIntensityLocation, GLuint ambientColorLocation,
+	                 GLuint diffuseIntensityLocation, GLuint directionLocation)
 {
 	// Ambient
 	glUniform3f(ambientColorLocation, color.x, color.y, color.z);
